add support for function expressions
